Refactor resolvers proxy methods
