+++ android/src/main/java/com/sicpa/didcomm/reactnative/DIDDocResolverProxy.kt
@@ -1,6 +1,6 @@
 package com.sicpa.didcomm.reactnative
 
-import android.util.Log
+import com.sicpa.didcomm.reactnative.utils.runBlockingWithTimeout
 import kotlinx.coroutines.*
 import org.didcommx.didcomm.diddoc.DIDDoc
 import org.didcommx.didcomm.diddoc.DIDDocResolver
@@ -18,18 +18,12 @@ class DIDDocResolverProxy(private val resolversProxyModule: ResolverProxyModule)
     override fun resolve(did: String): Optional<DIDDoc> {
         var resolvedDid: DIDDoc? = null
 
-        runBlocking {
-            try {
-                withTimeout(1500) {
-                    scope.launch {
-                        resolversProxyModule.sendEvent(ResolveDid(did))
-                        resolvedDid = resolvedDidChannel.receive()
-                    }.join()
-                }
-            } catch (e: TimeoutCancellationException) {
-                Log.e(TAG, "Resolve Did operation timed out")
-            }
+        val resolveDidJob = scope.launch {
+            resolversProxyModule.sendEvent(ResolveDid(did))
+            resolvedDid = resolvedDidChannel.receive()
         }
+
+        runBlockingWithTimeout(resolveDidJob, "${TAG}.resolve")
 
         return Optional.ofNullable(resolvedDid)
     }
