+++ android/src/main/java/com/sicpa/didcomm/reactnative/ResolverProxyModule.kt
@@ -5,6 +5,7 @@ import com.sicpa.didcomm.reactnative.model.*
 import com.facebook.react.bridge.*
 import com.facebook.react.module.annotations.ReactModule
 import com.facebook.react.modules.core.DeviceEventManagerModule
+import com.sicpa.didcomm.reactnative.utils.JsonUtils
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.channels.Channel
